optimize module download method
* install gcc7 if needed
* do not delete erverything in /usr/local/src anymore
* update module if needed

+++ nginx-build.sh
@@ -6,6 +6,7 @@ NGINX_STABLE=1.14.0
 NGINX_MAINLINE=$(curl -sL https://nginx.org/en/download.html 2>&1 | grep -E -o "nginx\\-[0-9.]+\\.tar[.a-z]*" | awk -F "nginx-" '/.tar.gz$/ {print $2}' | sed -e 's|.tar.gz||g' | head -n 1 2>&1)
 NAXSI_VER=0.56
 OPENSSL_VER=OpenSSL_1_1_1-pre8
+DIR_SRC=/usr/local/src
 
 # Colors
 CSI="\\033["
@@ -81,30 +82,32 @@ fi
 distro_version=$(lsb_release -sc)
 
 if [ "$distro_version" == "xenial" ]; then
-    echo -ne "       Installing gcc-7                       [..]\\r"
-    {
-        add-apt-repository ppa:jonathonf/gcc-7.1 -y
-        apt-get update
-        apt-get install gcc-7 g++-7  -y
-    } >> /tmp/nginx-ee.log 2>&1
-    
-    export CC="/usr/bin/gcc-7"
-    export CXX="/usr/bin/gc++-7"
-    if [ $? -eq 0 ]; then
-        echo -ne "       Installing gcc-7                       [${CGREEN}OK${CEND}]\\r"
-        echo -ne "\\n"
-    else
-        echo -e "        Installing gcc-7                      [${CRED}FAIL${CEND}]"
-        echo ""
-        echo "Please look at /tmp/nginx-ee.log"
-        echo ""
-        exit 1
+    if [ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-7_1-xenial.list ]; then
+        echo -ne "       Installing gcc-7                      [..]\\r"
+        {
+            add-apt-repository ppa:jonathonf/gcc-7.1 -y
+            apt-get update
+            apt-get install gcc-7 g++-7  -y
+        } >> /tmp/nginx-ee.log 2>&1
+        
+        export CC="/usr/bin/gcc-7"
+        export CXX="/usr/bin/gc++-7"
+        if [ $? -eq 0 ]; then
+            echo -ne "       Installing gcc-7                      [${CGREEN}OK${CEND}]\\r"
+            echo -ne "\\n"
+        else
+            echo -e "        Installing gcc-7                      [${CRED}FAIL${CEND}]"
+            echo ""
+            echo "Please look at /tmp/nginx-ee.log"
+            echo ""
+            exit 1
+        fi
     fi
 fi
 
 ## install prerequisites
 
-echo -ne "       Installing dependencies                [..]\\r"
+echo -ne "       Installing dependencies               [..]\\r"
 apt-get update >> /tmp/nginx-ee.log 2>&1
 apt-get install -y git build-essential libtool automake autoconf zlib1g-dev \
 libpcre3-dev libgd-dev libssl-dev libxslt1-dev libxml2-dev libgeoip-dev \
@@ -124,34 +127,94 @@ fi
 
 ## clean previous compilation
 
-rm -rf /usr/local/src/* >> /tmp/nginx-ee.log 2>&1
-cd /usr/local/src || exit
+cd $DIR_SRC || exit
+rm -rf ./*.tar.gz
 
 ## get additionals modules
 
 echo -ne "       Downloading additionals modules        [..]\\r"
 
 {
-    git clone https://github.com/FRiCKLE/ngx_cache_purge.git
-    git clone https://github.com/openresty/memc-nginx-module.git
-    git clone https://github.com/simpl/ngx_devel_kit.git
-    git clone https://github.com/openresty/headers-more-nginx-module.git
-    git clone https://github.com/openresty/echo-nginx-module.git
-    git clone https://github.com/yaoweibin/ngx_http_substitutions_filter_module.git
-    git clone https://github.com/openresty/redis2-nginx-module.git
-    git clone https://github.com/openresty/srcache-nginx-module.git
-    git clone https://github.com/openresty/set-misc-nginx-module.git
-    git clone https://github.com/sto/ngx_http_auth_pam_module.git
-    git clone https://github.com/vozlt/nginx-module-vts.git
+    if [ -d $DIR_SRC/ngx_cache_purge ]; then
+        { git -C $DIR_SRC/ngx_cache_purge pull origin master; }
+    else
+        { git clone https://github.com/FRiCKLE/ngx_cache_purge.git; }
+    fi
+    if [ -d $DIR_SRC/memc-nginx-module ]; then
+        { git -C $DIR_SRC/memc-nginx-module pull origin master; }
+    else
+        { git clone https://github.com/openresty/memc-nginx-module.git; }
+    fi
+    if [ -d $DIR_SRC/ngx_devel_kit ]; then
+        { git -C $DIR_SRC/ngx_devel_kit pull origin master; }
+    else
+        { git clone https://github.com/simpl/ngx_devel_kit.git; }
+    fi
+    if [ -d $DIR_SRC/headers-more-nginx-module ]; then
+        { git -C $DIR_SRC/headers-more-nginx-module pull origin master; }
+    else
+        { git clone https://github.com/openresty/headers-more-nginx-module.git; }
+    fi
+    if [ -d $DIR_SRC/echo-nginx-module ]; then
+        { git -C $DIR_SRC/echo-nginx-module pull origin master; }
+    else
+        { git clone https://github.com/openresty/echo-nginx-module.git; }
+    fi
+    if [ -d $DIR_SRC/echo-nginx-module ]; then
+        { git -C $DIR_SRC/echo-nginx-module pull origin master; }
+    else
+        { git clone https://github.com/openresty/echo-nginx-module.git; }
+    fi
+    if [ -d $DIR_SRC/ngx_http_substitutions_filter_module ]; then
+        { git -C $DIR_SRC/ngx_http_substitutions_filter_module pull origin master; }
+    else
+        { git clone https://github.com/yaoweibin/ngx_http_substitutions_filter_module.git; }
+    fi
+    if [ -d $DIR_SRC/redis2-nginx-module ]; then
+        { git -C $DIR_SRC/redis2-nginx-module pull origin master; }
+    else
+        { git clone https://github.com/openresty/redis2-nginx-module.git; }
+    fi
+    if [ -d $DIR_SRC/srcache-nginx-module ]; then
+        { git -C $DIR_SRC/srcache-nginx-module pull origin master; }
+    else
+        { git clone https://github.com/openresty/srcache-nginx-module.git; }
+    fi
+    if [ -d $DIR_SRC/set-misc-nginx-module ]; then
+        { git -C $DIR_SRC/set-misc-nginx-module pull origin master; }
+    else
+        { git clone https://github.com/openresty/set-misc-nginx-module.git; }
+    fi
+    if [ -d $DIR_SRC/ngx_http_auth_pam_module ]; then
+        { git -C $DIR_SRC/ngx_http_auth_pam_module pull origin master; }
+    else
+        { git clone https://github.com/sto/ngx_http_auth_pam_module.git; }
+    fi
+    if [ -d $DIR_SRC/nginx-module-vts ]; then
+        { git -C $DIR_SRC/nginx-module-vts pull origin master; }
+    else
+        { git clone https://github.com/vozlt/nginx-module-vts.git; }
+    fi
 } >> /tmp/nginx-ee.log 2>&1
 
-## ipsrcub module to anonymize user IP in nginx logs
-git clone https://github.com/masonicboom/ipscrub.git ipscrubtmp >> /tmp/nginx-ee.log 2>&1
-cp -rf /usr/local/src/ipscrubtmp/ipscrub /usr/local/src/ipscrub >> /tmp/nginx-ee.log 2>&1
+cd $DIR_SRC || exit
 
-wget https://people.freebsd.org/~osa/ngx_http_redis-0.3.8.tar.gz >> /tmp/nginx-ee.log 2>&1
-tar -zxf ngx_http_redis-0.3.8.tar.gz >> /tmp/nginx-ee.log 2>&1
-mv ngx_http_redis-0.3.8 ngx_http_redis
+## ipsrcub module to anonymize user IP in nginx logs
+{
+    rm -rf ipscrubtmp ipscrub
+    git clone https://github.com/masonicboom/ipscrub.git ipscrubtmp
+    cp -rf $DIR_SRC/ipscrubtmp/ipscrub $DIR_SRC/ipscrub
+} >> /tmp/nginx-ee.log 2>&1
+
+{
+    
+    cd $DIR_SRC || exit
+    
+    if [ ! -d $DIR_SRC/ngx_http_redis ]; then
+        wget https://people.freebsd.org/~osa/ngx_http_redis-0.3.8.tar.gz
+        mv ngx_http_redis-0.3.8 ngx_http_redis
+    fi
+} >> /tmp/nginx-ee.log 2>&1
 
 if [ $? -eq 0 ]; then
     echo -ne "       Downloading additionals modules        [${CGREEN}OK${CEND}]\\r"
@@ -166,11 +229,18 @@ fi
 
 # get brotli
 
-echo -ne "       Downloading brotli                     [..]\\r"
+cd $DIR_SRC || exit
 
-git clone https://github.com/google/ngx_brotli.git >> /tmp/nginx-ee.log 2>&1
-cd ngx_brotli || exit
-git submodule update --init --recursive >> /tmp/nginx-ee.log 2>&1
+echo -ne "       Downloading brotli                     [..]\\r"
+{
+    if [ -d $DIR_SRC/ngx_brotli ]; then
+        { git -C $DIR_SRC/ngx_brotli pull origin master; }
+    else
+        { git clone https://github.com/google/ngx_brotli.git; }
+    fi
+    cd ngx_brotli || exit
+    git submodule update --init --recursive
+} >> /tmp/nginx-ee.log 2>&1
 
 if [ $? -eq 0 ]; then
     echo -ne "       Downloading brotli                     [${CGREEN}OK${CEND}]\\r"
@@ -187,13 +257,19 @@ fi
 
 echo -ne "       Downloading openssl                    [..]\\r"
 
-cd /usr/local/src || exit
+cd $DIR_SRC || exit
+if [ -d $DIR_SRC/openssl ]
+then
+    cd $DIR_SRC/openssl || exit
+    git fetch >> /tmp/nginx-ee.log 2>&1
+    git checkout $OPENSSL_VER >> /tmp/nginx-ee.log 2>&1
+else
+    git clone https://github.com/openssl/openssl.git >> /tmp/nginx-ee.log 2>&1
+    cd $DIR_SRC/openssl || exit
+    git checkout $OPENSSL_VER >> /tmp/nginx-ee.log 2>&1
+fi
 
-git clone https://github.com/openssl/openssl.git >> /tmp/nginx-ee.log 2>&1
-cd openssl || exit
-git checkout $OPENSSL_VER >> /tmp/nginx-ee.log 2>&1
 
-cd /usr/local/src || exit
 
 if [ $? -eq 0 ]; then
     echo -ne "       Downloading openssl                    [${CGREEN}OK${CEND}]\\r"
@@ -207,14 +283,17 @@ else
 fi
 
 ## get naxsi
-
+cd $DIR_SRC || exit
 if [ "$naxsi" = "y" ]
 then
     echo -ne "       Downloading naxsi                      [..]\\r"
+    if [ -d $DIR_SRC/naxsi ]; then
+        rm -rf $DIR_SRC/naxsi
+    fi
     wget -O naxsi.tar.gz https://github.com/nbs-system/naxsi/archive/$NAXSI_VER.tar.gz >> /tmp/nginx-ee.log 2>&1
     tar xvzf naxsi.tar.gz >> /tmp/nginx-ee.log 2>&1
     mv naxsi-$NAXSI_VER naxsi
-    cd /usr/local/src || exit
+    
     
     if [ $? -eq 0 ]; then
         echo -ne "       Downloading naxsi                      [${CGREEN}OK${CEND}]\\r"
@@ -231,16 +310,17 @@ fi
 
 
 ## get ngx_pagespeed
-
+cd $DIR_SRC || exit
 if [ "$pagespeed" = "y" ]
 then
     echo -ne "       Downloading pagespeed                  [..]\\r"
     {
+        rm -rf incubator-pagespeed-ngx-latest-beta install
         wget https://ngxpagespeed.com/install
         chmod +x install
-        ./install --ngx-pagespeed-version latest-beta -b /usr/local/src
+        ./install --ngx-pagespeed-version latest-beta -b $DIR_SRC
     } >> /tmp/nginx-ee.log 2>&1
-    cd /usr/local/src/ || exit
+    
     
     if [ $? -eq 0 ]; then
         echo -ne "       Downloading pagespeed                  [${CGREEN}OK${CEND}]\\r"
@@ -255,13 +335,16 @@ then
 fi
 
 ## get nginx
-
+cd $DIR_SRC || exit
 echo -ne "       Downloading nginx                      [..]\\r"
+if [ -d $DIR_SRC/nginx ]; then
+    rm -rf $DIR_SRC/nginx nginx-*
+fi
 wget http://nginx.org/download/nginx-${NGINX_RELEASE}.tar.gz >> /tmp/nginx-ee.log 2>&1
 tar -xzvf nginx-${NGINX_RELEASE}.tar.gz >> /tmp/nginx-ee.log 2>&1
 mv nginx-${NGINX_RELEASE} nginx
 
-cd /usr/local/src/nginx/ || exit
+cd $DIR_SRC/nginx/ || exit
 
 if [ $? -eq 0 ]; then
     echo -ne "       Downloading nginx                      [${CGREEN}OK${CEND}]\\r"
@@ -379,8 +462,8 @@ fi
 
 {
     systemctl unmask nginx
-    systemctl enable nginx
-    systemctl start nginx
+    systemctl enable nginx.service
+    systemctl start nginx.service
     apt-mark hold nginx-ee nginx-common
     systemctl restart nginx
     nginx -t
